allegedly minimizing the writing editor
